requirements.txt
requests
bs4
fake_useragent
tenacity
mysqlclient
mysql-connector-python
